+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixed
 
 ### Changed
+- Console wordwrap: cut long words
 - Symfony Console 3.1
 - Symfony Translation 3.1
 - PHPUnit 5.5